backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Image/Video processing
opencv-python==4.8.1.78
Pillow==10.1.0
imageio==2.33.0
numpy==1.24.3

# PyTorch and ML models
torch==2.1.0
torchvision==0.16.0

# FaceForensics++ Model dependencies
timm==0.9.12
facenet-pytorch==2.5.3

# Optional: HuggingFace transformers (if you want to keep the old model as fallback)
transformers==4.35.2

# Development
python-dotenv==1.0.0